Shell sorting

+++ 2025_10_29_my_shell_sort/io.cpp
@@ -3,9 +3,9 @@
 #include "io.hpp"
 
 namespace gpi {
-    void print_array(const char* comment, const int* array, size_t array_size) {
+    void print_array(const char* comment, const int* array, int array_size) {
         std::cout << comment;
-        for (size_t i = 0; i < array_size; ++i) {
+        for (int i = 0; i < array_size; ++i) {
             std::cout << array[i] << ' ';
         }
         std::cout << std::endl;
