+++ 2025_10_29_my_shell_sort/io.hpp
@@ -4,7 +4,7 @@
 #include <cstddef>
 
 namespace gpi {
-    void print_array(const char* comment, const int* array, size_t array_size);
+    void print_array(const char* comment, const int* array, int array_size);
 }
 
 #endif
