+++ 2025_10_29_my_shell_sort/shell_sorting.hpp
@@ -4,9 +4,9 @@
 #include <cstddef>
 
 namespace gpi {
-    size_t* generate_gaps_sequence(size_t array_size, size_t& sequence_size);
-    void shell_sort(int* array, size_t array_size);
-    void fill_random_values(int* array, size_t array_size);
+    int* generate_gaps_sequence(int array_size, int& sequence_size);
+    void shell_sort(int* array, int array_size);
+    void fill_random_values(int* array, int array_size);
 }
 
 #endif
